add: livewire top torrents panel
By only showing the newest torrents by default, we can abuse the fact that it's the only tab we can easily put an index on to make the home page load fast, and keep the 4 other tabs that take ~1 second to query lazily loaded until clicked upon.
