+++ app/Http/Controllers/HomeController.php
@@ -18,12 +18,10 @@ use App\Models\Bookmark;
 use App\Models\FeaturedTorrent;
 use App\Models\FreeleechToken;
 use App\Models\Group;
-use App\Models\Movie;
 use App\Models\Poll;
 use App\Models\Post;
 use App\Models\Topic;
 use App\Models\Torrent;
-use App\Models\Tv;
 use App\Models\User;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Http\Request;
@@ -86,420 +84,7 @@ class HomeController extends Controller
                     ->get()
             ),
             'articles' => $articles,
-            'newest'   => cache()->remember(
-                'newest_torrents',
-                $expiresAt,
-                function () use ($user) {
-                    $newest = Torrent::query()
-                        ->select([
-                            'id',
-                            'name',
-                            'user_id',
-                            'category_id',
-                            'type_id',
-                            'resolution_id',
-                            'tmdb',
-                            'igdb',
-                            'size',
-                            'seeders',
-                            'leechers',
-                            'times_completed',
-                            'created_at'
-                        ])
-                        ->with(['user.group', 'category', 'type', 'resolution'])
-                        ->withExists([
-                            'bookmarks'          => fn ($query) => $query->where('user_id', '=', $user->id),
-                            'freeleechTokens'    => fn ($query) => $query->where('user_id', '=', $user->id),
-                            'history as seeding' => fn ($query) => $query->where('user_id', '=', $user->id)
-                                ->where('active', '=', 1)
-                                ->where('seeder', '=', 1),
-                            'history as leeching' => fn ($query) => $query->where('user_id', '=', $user->id)
-                                ->where('active', '=', 1)
-                                ->where('seeder', '=', 0),
-                            'history as not_completed' => fn ($query) => $query->where('user_id', '=', $user->id)
-                                ->where('active', '=', 0)
-                                ->where('seeder', '=', 0)
-                                ->whereNull('completed_at'),
-                            'history as not_seeding' => fn ($query) => $query->where('user_id', '=', $user->id)
-                                ->where('active', '=', 0)
-                                ->where(
-                                    fn ($query) => $query
-                                        ->where('seeder', '=', 1)
-                                        ->orWhereNotNull('completed_at')
-                                ),
-                        ])
-                        ->selectRaw("
-                    CASE
-                        WHEN category_id IN (SELECT `id` from `categories` where `movie_meta` = 1) THEN 'movie'
-                        WHEN category_id IN (SELECT `id` from `categories` where `tv_meta` = 1) THEN 'tv'
-                        WHEN category_id IN (SELECT `id` from `categories` where `game_meta` = 1) THEN 'game'
-                        WHEN category_id IN (SELECT `id` from `categories` where `music_meta` = 1) THEN 'music'
-                        WHEN category_id IN (SELECT `id` from `categories` where `no_meta` = 1) THEN 'no'
-                    END as meta
-                ")
-                        ->withCount(['thanks', 'comments'])
-                        ->orderByDesc('id')
-                        ->take(5)
-                        ->get();
-
-                    $movieIds = $newest->where('meta', '=', 'movie')->pluck('tmdb');
-                    $tvIds = $newest->where('meta', '=', 'tv')->pluck('tmdb');
-                    $gameIds = $newest->where('meta', '=', 'game')->pluck('igdb');
-
-                    $movies = Movie::with('genres')->whereIntegerInRaw('id', $movieIds)->get()->keyBy('id');
-                    $tv = Tv::with('genres')->whereIntegerInRaw('id', $tvIds)->get()->keyBy('id');
-                    $games = [];
-
-                    foreach ($gameIds as $gameId) {
-                        $games[$gameId] = \MarcReichel\IGDBLaravel\Models\Game::with(['cover' => ['url', 'image_id']])->find($gameId);
-                    }
-
-                    $newest = $newest->map(function ($torrent) use ($movies, $tv, $games) {
-                        $torrent->meta = match ($torrent->meta) {
-                            'movie' => $movies[$torrent->tmdb] ?? null,
-                            'tv'    => $tv[$torrent->tmdb] ?? null,
-                            'game'  => $games[$torrent->igdb] ?? null,
-                            default => null,
-                        };
-
-                        return $torrent;
-                    });
-
-                    return $newest;
-                }
-            ),
-            'seeded' => cache()->remember(
-                'seeded_torrents',
-                $expiresAt,
-                function () use ($user) {
-                    $seeded = Torrent::query()
-                        ->select([
-                            'id',
-                            'name',
-                            'user_id',
-                            'category_id',
-                            'type_id',
-                            'resolution_id',
-                            'tmdb',
-                            'igdb',
-                            'size',
-                            'seeders',
-                            'leechers',
-                            'times_completed',
-                            'created_at'
-                        ])
-                        ->with(['user.group', 'category', 'type', 'resolution'])
-                        ->withExists([
-                            'bookmarks'          => fn ($query) => $query->where('user_id', '=', $user->id),
-                            'freeleechTokens'    => fn ($query) => $query->where('user_id', '=', $user->id),
-                            'history as seeding' => fn ($query) => $query->where('user_id', '=', $user->id)
-                                ->where('active', '=', 1)
-                                ->where('seeder', '=', 1),
-                            'history as leeching' => fn ($query) => $query->where('user_id', '=', $user->id)
-                                ->where('active', '=', 1)
-                                ->where('seeder', '=', 0),
-                            'history as not_completed' => fn ($query) => $query->where('user_id', '=', $user->id)
-                                ->where('active', '=', 0)
-                                ->where('seeder', '=', 0)
-                                ->whereNull('completed_at'),
-                            'history as not_seeding' => fn ($query) => $query->where('user_id', '=', $user->id)
-                                ->where('active', '=', 0)
-                                ->where(
-                                    fn ($query) => $query
-                                        ->where('seeder', '=', 1)
-                                        ->orWhereNotNull('completed_at')
-                                ),
-                        ])
-                        ->selectRaw("
-                    CASE
-                        WHEN category_id IN (SELECT `id` from `categories` where `movie_meta` = 1) THEN 'movie'
-                        WHEN category_id IN (SELECT `id` from `categories` where `tv_meta` = 1) THEN 'tv'
-                        WHEN category_id IN (SELECT `id` from `categories` where `game_meta` = 1) THEN 'game'
-                        WHEN category_id IN (SELECT `id` from `categories` where `music_meta` = 1) THEN 'music'
-                        WHEN category_id IN (SELECT `id` from `categories` where `no_meta` = 1) THEN 'no'
-                    END as meta
-                ")
-                        ->withCount(['thanks', 'comments'])
-                        ->orderByDesc('seeders')
-                        ->take(5)
-                        ->get();
-
-                    $movieIds = $seeded->where('meta', '=', 'movie')->pluck('tmdb');
-                    $tvIds = $seeded->where('meta', '=', 'tv')->pluck('tmdb');
-                    $gameIds = $seeded->where('meta', '=', 'game')->pluck('igdb');
-
-                    $movies = Movie::with('genres')->whereIntegerInRaw('id', $movieIds)->get()->keyBy('id');
-                    $tv = Tv::with('genres')->whereIntegerInRaw('id', $tvIds)->get()->keyBy('id');
-                    $games = [];
-
-                    foreach ($gameIds as $gameId) {
-                        $games[$gameId] = \MarcReichel\IGDBLaravel\Models\Game::with(['cover' => ['url', 'image_id']])->find($gameId);
-                    }
-
-                    $seeded = $seeded->map(function ($torrent) use ($movies, $tv, $games) {
-                        $torrent->meta = match ($torrent->meta) {
-                            'movie' => $movies[$torrent->tmdb] ?? null,
-                            'tv'    => $tv[$torrent->tmdb] ?? null,
-                            'game'  => $games[$torrent->igdb] ?? null,
-                            default => null,
-                        };
-
-                        return $torrent;
-                    });
-
-                    return $seeded;
-                }
-            ),
-            'dying' => cache()->remember(
-                'dying_torrents',
-                $expiresAt,
-                function () use ($user) {
-                    $dying = Torrent::query()
-                        ->select([
-                            'id',
-                            'name',
-                            'user_id',
-                            'category_id',
-                            'type_id',
-                            'resolution_id',
-                            'tmdb',
-                            'igdb',
-                            'size',
-                            'seeders',
-                            'leechers',
-                            'times_completed',
-                            'created_at'
-                        ])
-                        ->with(['user.group', 'category', 'type', 'resolution'])
-                        ->withExists([
-                            'bookmarks'          => fn ($query) => $query->where('user_id', '=', $user->id),
-                            'freeleechTokens'    => fn ($query) => $query->where('user_id', '=', $user->id),
-                            'history as seeding' => fn ($query) => $query->where('user_id', '=', $user->id)
-                                ->where('active', '=', 1)
-                                ->where('seeder', '=', 1),
-                            'history as leeching' => fn ($query) => $query->where('user_id', '=', $user->id)
-                                ->where('active', '=', 1)
-                                ->where('seeder', '=', 0),
-                            'history as not_completed' => fn ($query) => $query->where('user_id', '=', $user->id)
-                                ->where('active', '=', 0)
-                                ->where('seeder', '=', 0)
-                                ->whereNull('completed_at'),
-                            'history as not_seeding' => fn ($query) => $query->where('user_id', '=', $user->id)
-                                ->where('active', '=', 0)
-                                ->where(
-                                    fn ($query) => $query
-                                        ->where('seeder', '=', 1)
-                                        ->orWhereNotNull('completed_at')
-                                ),
-                        ])
-                        ->selectRaw("
-                    CASE
-                        WHEN category_id IN (SELECT `id` from `categories` where `movie_meta` = 1) THEN 'movie'
-                        WHEN category_id IN (SELECT `id` from `categories` where `tv_meta` = 1) THEN 'tv'
-                        WHEN category_id IN (SELECT `id` from `categories` where `game_meta` = 1) THEN 'game'
-                        WHEN category_id IN (SELECT `id` from `categories` where `music_meta` = 1) THEN 'music'
-                        WHEN category_id IN (SELECT `id` from `categories` where `no_meta` = 1) THEN 'no'
-                    END as meta
-                ")
-                        ->withCount(['thanks', 'comments'])
-                        ->where('seeders', '=', 1)
-                        ->where('times_completed', '>=', 1)
-                        ->orderByDesc('leechers')
-                        ->take(5)
-                        ->get();
-
-                    $movieIds = $dying->where('meta', '=', 'movie')->pluck('tmdb');
-                    $tvIds = $dying->where('meta', '=', 'tv')->pluck('tmdb');
-                    $gameIds = $dying->where('meta', '=', 'game')->pluck('igdb');
-
-                    $movies = Movie::with('genres')->whereIntegerInRaw('id', $movieIds)->get()->keyBy('id');
-                    $tv = Tv::with('genres')->whereIntegerInRaw('id', $tvIds)->get()->keyBy('id');
-                    $games = [];
-
-                    foreach ($gameIds as $gameId) {
-                        $games[$gameId] = \MarcReichel\IGDBLaravel\Models\Game::with(['cover' => ['url', 'image_id']])->find($gameId);
-                    }
-
-                    $dying = $dying->map(function ($torrent) use ($movies, $tv, $games) {
-                        $torrent->meta = match ($torrent->meta) {
-                            'movie' => $movies[$torrent->tmdb] ?? null,
-                            'tv'    => $tv[$torrent->tmdb] ?? null,
-                            'game'  => $games[$torrent->igdb] ?? null,
-                            default => null,
-                        };
-
-                        return $torrent;
-                    });
-
-                    return $dying;
-                }
-            ),
-            'leeched' => cache()->remember(
-                'leeched_torrents',
-                $expiresAt,
-                function () use ($user) {
-                    $leeched = Torrent::query()
-                        ->select([
-                            'id',
-                            'name',
-                            'user_id',
-                            'category_id',
-                            'type_id',
-                            'resolution_id',
-                            'tmdb',
-                            'igdb',
-                            'size',
-                            'seeders',
-                            'leechers',
-                            'times_completed',
-                            'created_at'
-                        ])
-                        ->with(['user.group', 'category', 'type', 'resolution'])
-                        ->withExists([
-                            'bookmarks'          => fn ($query) => $query->where('user_id', '=', $user->id),
-                            'freeleechTokens'    => fn ($query) => $query->where('user_id', '=', $user->id),
-                            'history as seeding' => fn ($query) => $query->where('user_id', '=', $user->id)
-                                ->where('active', '=', 1)
-                                ->where('seeder', '=', 1),
-                            'history as leeching' => fn ($query) => $query->where('user_id', '=', $user->id)
-                                ->where('active', '=', 1)
-                                ->where('seeder', '=', 0),
-                            'history as not_completed' => fn ($query) => $query->where('user_id', '=', $user->id)
-                                ->where('active', '=', 0)
-                                ->where('seeder', '=', 0)
-                                ->whereNull('completed_at'),
-                            'history as not_seeding' => fn ($query) => $query->where('user_id', '=', $user->id)
-                                ->where('active', '=', 0)
-                                ->where(
-                                    fn ($query) => $query
-                                        ->where('seeder', '=', 1)
-                                        ->orWhereNotNull('completed_at')
-                                ),
-                        ])
-                        ->selectRaw("
-                    CASE
-                        WHEN category_id IN (SELECT `id` from `categories` where `movie_meta` = 1) THEN 'movie'
-                        WHEN category_id IN (SELECT `id` from `categories` where `tv_meta` = 1) THEN 'tv'
-                        WHEN category_id IN (SELECT `id` from `categories` where `game_meta` = 1) THEN 'game'
-                        WHEN category_id IN (SELECT `id` from `categories` where `music_meta` = 1) THEN 'music'
-                        WHEN category_id IN (SELECT `id` from `categories` where `no_meta` = 1) THEN 'no'
-                    END as meta
-                ")
-                        ->withCount(['thanks', 'comments'])
-                        ->orderByDesc('leechers')
-                        ->take(5)
-                        ->get();
-
-                    $movieIds = $leeched->where('meta', '=', 'movie')->pluck('tmdb');
-                    $tvIds = $leeched->where('meta', '=', 'tv')->pluck('tmdb');
-                    $gameIds = $leeched->where('meta', '=', 'game')->pluck('igdb');
-
-                    $movies = Movie::with('genres')->whereIntegerInRaw('id', $movieIds)->get()->keyBy('id');
-                    $tv = Tv::with('genres')->whereIntegerInRaw('id', $tvIds)->get()->keyBy('id');
-                    $games = [];
-
-                    foreach ($gameIds as $gameId) {
-                        $games[$gameId] = \MarcReichel\IGDBLaravel\Models\Game::with(['cover' => ['url', 'image_id']])->find($gameId);
-                    }
-
-                    $leeched = $leeched->map(function ($torrent) use ($movies, $tv, $games) {
-                        $torrent->meta = match ($torrent->meta) {
-                            'movie' => $movies[$torrent->tmdb] ?? null,
-                            'tv'    => $tv[$torrent->tmdb] ?? null,
-                            'game'  => $games[$torrent->igdb] ?? null,
-                            default => null,
-                        };
-
-                        return $torrent;
-                    });
-
-                    return $leeched;
-                }
-            ),
-            'dead' => cache()->remember(
-                'dead_torrents',
-                $expiresAt,
-                function () use ($user) {
-                    $dead = Torrent::query()
-                        ->select([
-                            'id',
-                            'name',
-                            'user_id',
-                            'category_id',
-                            'type_id',
-                            'resolution_id',
-                            'tmdb',
-                            'igdb',
-                            'size',
-                            'seeders',
-                            'leechers',
-                            'times_completed',
-                            'created_at'
-                        ])
-                        ->with(['user.group', 'category', 'type', 'resolution'])
-                        ->withExists([
-                            'bookmarks'          => fn ($query) => $query->where('user_id', '=', $user->id),
-                            'freeleechTokens'    => fn ($query) => $query->where('user_id', '=', $user->id),
-                            'history as seeding' => fn ($query) => $query->where('user_id', '=', $user->id)
-                                ->where('active', '=', 1)
-                                ->where('seeder', '=', 1),
-                            'history as leeching' => fn ($query) => $query->where('user_id', '=', $user->id)
-                                ->where('active', '=', 1)
-                                ->where('seeder', '=', 0),
-                            'history as not_completed' => fn ($query) => $query->where('user_id', '=', $user->id)
-                                ->where('active', '=', 0)
-                                ->where('seeder', '=', 0)
-                                ->whereNull('completed_at'),
-                            'history as not_seeding' => fn ($query) => $query->where('user_id', '=', $user->id)
-                                ->where('active', '=', 0)
-                                ->where(
-                                    fn ($query) => $query
-                                        ->where('seeder', '=', 1)
-                                        ->orWhereNotNull('completed_at')
-                                ),
-                        ])
-                        ->selectRaw("
-                    CASE
-                        WHEN category_id IN (SELECT `id` from `categories` where `movie_meta` = 1) THEN 'movie'
-                        WHEN category_id IN (SELECT `id` from `categories` where `tv_meta` = 1) THEN 'tv'
-                        WHEN category_id IN (SELECT `id` from `categories` where `game_meta` = 1) THEN 'game'
-                        WHEN category_id IN (SELECT `id` from `categories` where `music_meta` = 1) THEN 'music'
-                        WHEN category_id IN (SELECT `id` from `categories` where `no_meta` = 1) THEN 'no'
-                    END as meta
-                ")
-                        ->withCount(['thanks', 'comments'])
-                        ->where('seeders', '=', 0)
-                        ->orderByDesc('leechers')
-                        ->take(5)
-                        ->get();
-
-                    $movieIds = $dead->where('meta', '=', 'movie')->pluck('tmdb');
-                    $tvIds = $dead->where('meta', '=', 'tv')->pluck('tmdb');
-                    $gameIds = $dead->where('meta', '=', 'game')->pluck('igdb');
-
-                    $movies = Movie::with('genres')->whereIntegerInRaw('id', $movieIds)->get()->keyBy('id');
-                    $tv = Tv::with('genres')->whereIntegerInRaw('id', $tvIds)->get()->keyBy('id');
-                    $games = [];
-
-                    foreach ($gameIds as $gameId) {
-                        $games[$gameId] = \MarcReichel\IGDBLaravel\Models\Game::with(['cover' => ['url', 'image_id']])->find($gameId);
-                    }
-
-                    $dead = $dead->map(function ($torrent) use ($movies, $tv, $games) {
-                        $torrent->meta = match ($torrent->meta) {
-                            'movie' => $movies[$torrent->tmdb] ?? null,
-                            'tv'    => $tv[$torrent->tmdb] ?? null,
-                            'game'  => $games[$torrent->igdb] ?? null,
-                            default => null,
-                        };
-
-                        return $torrent;
-                    });
-
-                    return $dead;
-                }
-            ),
-            'topics' => cache()->remember(
+            'topics'   => cache()->remember(
                 'latest_topics:by-group:'.auth()->user()->group_id,
                 $expiresAt,
                 fn () => Topic::query()
