+++ phpstan-baseline.neon
@@ -641,11 +641,6 @@ parameters:
 			path: app/Http/Controllers/ApprovedRequestFillController.php
 
 		-
-			message: "#^Access to an undefined property App\\\\Models\\\\Torrent\\:\\:\\$meta\\.$#"
-			count: 10
-			path: app/Http/Controllers/HomeController.php
-
-		-
 			message: "#^Cannot access property \\$alive on bool\\|stdClass\\.$#"
 			count: 1
 			path: app/Http/Controllers/RssController.php
